+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Unreleased
 
 - Most of the `PrimitiveWetModel` is now Reactant compatible, this includes a revision to make most structs fully parametric [#985](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/985)
+- Fix GPU inverse Legendre transform for single-layer spectral fields on GPU [#1173](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1173)
+- Revised extended differentiability tests [#1167](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1167)
+- Compatability with Terrarium 0.1.4 [#1157](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1157) [#1170](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1170)
+- Add time chunk buffering for ZarrOutput [#1162](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1162)
+- Enable nested AD on CPU by shortening type name of `SpectralTransform`, e.g. by converting the gradient arrays into a type `Gradient` instead of a `NamedTuple` [#1165](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1165) [#1166](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1165)
+- Forward mode differentiation of the whole model now works (in 1.10) [#1164](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1164)
+- Enzyme Forward rules for SpectralTransform [#767](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/767)
 - Revised EnzymeRules for SpeedyTransforms [#1151](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1151)
 - Eliminate dynamic dispatchs in the model integration code and unit test for them with JET [#1151](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1151)
 - ZarrOutput ensemble mode hardened for parallel jobs, Zarr store metadata consolidated at creation for faster opening [#1160](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1160)